#138 리팩토링 후 기능 확인 완료
