--- EUMMEYO/View/Calendar/ChildView/AddMemoViewModel.swift
@@ -1,16 +1,0 @@
-//
-//  AddMemoViewModel.swift
-//  EUMMEYO
-//
-//  Created by 김동현 on 2/11/25.
-//
-
-import SwiftUI
-import Combine
-
-final class AddMemoViewModel: ObservableObject {
-    var cancellables = Set<AnyCancellable>()
-    
-    
-    
-}
